backend/requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
python-socketio==5.12.1
motor==3.6.1
pydantic==2.10.5
python-dotenv==1.0.1
websockets==14.2


